requirements.txt
scikit-learn
matplotlib
seaborn
pandas
numpy
joblib
fastapi
uvicorn
streamlit
pillow
